chore(storybook): add branding
